+++ src/libfetchers/registry.cc
@@ -1,7 +1,7 @@
 #include "nix/fetchers/fetch-settings.hh"
 #include "nix/fetchers/registry.hh"
 #include "nix/fetchers/tarball.hh"
-#include "nix/store/filetransfer.hh"
+#include "nix/util/filetransfer.hh"
 #include "nix/util/users.hh"
 #include "nix/store/store-api.hh"
 #include "nix/store/local-fs-store.hh"
