+++ src/libfetchers/tarball.cc
@@ -1,7 +1,7 @@
 #include "nix/fetchers/tarball.hh"
 #include "nix/fetchers/fetchers.hh"
 #include "nix/fetchers/cache.hh"
-#include "nix/store/filetransfer.hh"
+#include "nix/util/filetransfer.hh"
 #include "nix/store/store-api.hh"
 #include "nix/util/archive.hh"
 #include "nix/util/tarfile.hh"
