+++ src/libstore-tests/meson.build
@@ -64,8 +64,6 @@ sources = files(
   'derived-path.cc',
   'downstream-placeholder.cc',
   'dummy-store.cc',
-  'filetransfer-request.cc',
-  'filetransfer-retry.cc',
   'http-binary-cache-store.cc',
   'legacy-ssh-store.cc',
   'local-binary-cache-store.cc',
@@ -84,7 +82,6 @@ sources = files(
   'realisation.cc',
   'references.cc',
   's3-binary-cache-store.cc',
-  's3-url.cc',
   'serve-protocol.cc',
   'ssh-store.cc',
   'store-open.cc',
