+++ src/libstore/builtins/fetchurl.cc
@@ -1,5 +1,5 @@
 #include "nix/store/builtins.hh"
-#include "nix/store/filetransfer.hh"
+#include "nix/util/filetransfer.hh"
 #include "nix/store/store-api.hh"
 #include "nix/util/archive.hh"
 #include "nix/util/compression.hh"
