+++ src/libstore/globals.cc
@@ -8,14 +8,13 @@
 #include "nix/util/abstract-setting-to-json.hh"
 #include "nix/util/compute-levels.hh"
 #include "nix/util/executable-path.hh"
-#include "nix/store/filetransfer.hh"
+#include "nix/util/filetransfer.hh"
 
 #include <algorithm>
 #include <map>
 #include <mutex>
 #include <thread>
 
-#include <curl/curl.h>
 #include <nlohmann/json.hpp>
 
 #include <limits.h>
@@ -134,51 +133,6 @@ void loadConfFile(AbstractConfig & config)
     if (nixConfEnv.has_value()) {
         config.applyConfig(nixConfEnv.value(), "NIX_CONFIG");
     }
-}
-
-/**
- * On Windows, NIX_CONF_DIR (and other directories like NIX_STATE_DIR, NIX_LOG_DIR)
- * are not defined at compile time, so we determine paths at runtime using the
- * Windows known folders API (FOLDERID_ProgramData). This allows Nix to work
- * correctly regardless of which drive Windows is installed on.
- */
-const std::filesystem::path & nixConfDir()
-{
-    static const std::filesystem::path dir = getEnvOsNonEmpty(OS_STR("NIX_CONF_DIR"))
-                                                 .transform([](auto && s) { return std::filesystem::path(s); })
-                                                 .or_else([]() -> std::optional<std::filesystem::path> {
-#ifdef _WIN32
-#  ifdef NIX_CONF_DIR
-#    error "NIX_CONF_DIR should not be defined on Windows"
-#  endif
-                                                     return windows::known_folders::getProgramData() / "nix" / "conf";
-#else
-                                                     return NIX_CONF_DIR;
-#endif
-                                                 })
-                                                 .transform([](auto && s) { return canonPath(s); })
-                                                 .value();
-    return dir;
-}
-
-const std::vector<std::filesystem::path> & nixUserConfFiles()
-{
-    static const std::vector<std::filesystem::path> files = [] {
-        // Use the paths specified in NIX_USER_CONF_FILES if it has been defined
-        auto nixConfFiles = getEnvOs(OS_STR("NIX_USER_CONF_FILES"));
-        if (nixConfFiles.has_value()) {
-            return ExecutablePath::parse(*nixConfFiles).directories;
-        }
-
-        // Use the paths specified by the XDG spec
-        std::vector<std::filesystem::path> files;
-        auto dirs = getConfigDirs();
-        for (auto & dir : dirs) {
-            files.insert(files.end(), dir / "nix.conf");
-        }
-        return files;
-    }();
-    return files;
 }
 
 unsigned int Settings::getDefaultCores()
@@ -310,10 +264,6 @@ ProfileDirsOptions Settings::getProfileDirsOptions() const
         .useXDGBaseDirectories = useXDGBaseDirectories,
     };
 }
-
-std::string nixVersion = PACKAGE_VERSION;
-
-const std::string determinateNixVersion = DETERMINATE_NIX_VERSION;
 
 NLOHMANN_JSON_SERIALIZE_ENUM(
     SandboxMode,
@@ -647,19 +597,6 @@ void initLibStore(bool loadConfig)
         loadConfFile(globalConfig);
 
     preloadNSS();
-
-    /* Because of an objc quirk[1], calling curl_global_init for the first time
-       after fork() will always result in a crash.
-       Up until now the solution has been to set OBJC_DISABLE_INITIALIZE_FORK_SAFETY
-       for every nix process to ignore that error.
-       Instead of working around that error we address it at the core -
-       by calling curl_global_init here, which should mean curl will already
-       have been initialized by the time we try to do so in a forked process.
-
-       [1]
-       https://github.com/apple-oss-distributions/objc4/blob/01edf1705fbc3ff78a423cd21e03dfc21eb4d780/runtime/objc-initialize.mm#L614-L636
-    */
-    curl_global_init(CURL_GLOBAL_ALL);
 #ifdef __APPLE__
     /* On macOS, don't use the per-session TMPDIR (as set e.g. by
        sshd). This breaks build users because they don't have access
