+++ src/libstore/http-binary-cache-store.cc
@@ -1,5 +1,5 @@
 #include "nix/store/http-binary-cache-store.hh"
-#include "nix/store/filetransfer.hh"
+#include "nix/util/filetransfer.hh"
 #include "nix/store/globals.hh"
 #include "nix/store/nar-info-disk-cache.hh"
 #include "nix/store/sqlite.hh"
