+++ src/libstore/include/nix/store/meson.build
@@ -12,7 +12,6 @@ config_pub_h = configure_file(
 headers = [ config_pub_h ] + files(
   'active-builds.hh',
   'async-path-writer.hh',
-  'aws-creds.hh',
   'binary-cache-store.hh',
   'build-result.hh',
   'build/build-log.hh',
@@ -42,10 +41,7 @@ headers = [ config_pub_h ] + files(
   'dummy-store-impl.hh',
   'dummy-store.hh',
   'export-import.hh',
-  'filetransfer-impl.hh',
-  'filetransfer.hh',
   'gc-store.hh',
-  'global-paths.hh',
   'globals.hh',
   'http-binary-cache-store.hh',
   'indirect-root-store.hh',
@@ -83,7 +79,6 @@ headers = [ config_pub_h ] + files(
   'remote-store.hh',
   'restricted-store.hh',
   's3-binary-cache-store.hh',
-  's3-url.hh',
   'serve-protocol-connection.hh',
   'serve-protocol-impl.hh',
   'serve-protocol.hh',
