+++ src/libutil-tests/meson.build
@@ -73,6 +73,8 @@ sources = files(
   'file-descriptor.cc',
   'file-system-at.cc',
   'file-system.cc',
+  'filetransfer-request.cc',
+  'filetransfer-retry.cc',
   'fun.cc',
   'git.cc',
   'hash.cc',
@@ -91,6 +93,7 @@ sources = files(
   'position.cc',
   'processes.cc',
   'ref.cc',
+  's3-url.cc',
   'serialise.cc',
   'sort.cc',
   'source-accessor.cc',
