+++ src/libutil/include/nix/util/global-paths.hh
