+++ src/libutil/include/nix/util/meson.build
@@ -16,6 +16,7 @@ headers = [ config_pub_h ] + files(
   'args/root.hh',
   'array-from-string-literal.hh',
   'async.hh',
+  'aws-creds.hh',
   'base-n.hh',
   'base-nix-32.hh',
   'bump-memory-resource.hh',
@@ -48,12 +49,15 @@ headers = [ config_pub_h ] + files(
   'file-path.hh',
   'file-system-at.hh',
   'file-system.hh',
+  'filetransfer-impl.hh',
+  'filetransfer.hh',
   'finally.hh',
   'fmt.hh',
   'forwarding-source-accessor.hh',
   'fs-sink.hh',
   'fun.hh',
   'git.hh',
+  'global-paths.hh',
   'hash.hh',
   'hilite.hh',
   'json-impls.hh',
@@ -80,6 +84,7 @@ headers = [ config_pub_h ] + files(
   'ref.hh',
   'regex-combinators.hh',
   'repair-flag.hh',
+  's3-url.hh',
   'sentry.hh',
   'serialise.hh',
   'signals.hh',
@@ -107,5 +112,6 @@ headers = [ config_pub_h ] + files(
   'users.hh',
   'util.hh',
   'variant-wrapper.hh',
+  'version.hh',
   'xml-writer.hh',
 )
