+++ src/libutil/include/nix/util/s3-url.hh
@@ -1,6 +1,6 @@
 #pragma once
 ///@file
-#include "nix/store/config.hh"
+#include "nix/util/config.hh"
 #include "nix/util/error.hh"
 #include "nix/util/url.hh"
 #include "nix/util/util.hh"
